refactor: AdminIdProvider 구성 변경 및 adminId 삽입 로직 위치 변경
